Ajout de la construction de l'image docker et correction de la fonction select

+++ src/test/java/Parser_DataFrameTest.java
@@ -83,7 +83,7 @@ class Parser_DataFrameTest {
                         values[i][j] = randomString(random.nextInt(10) + 5);
                         break;
                     case 2:
-                        values[i][j] = random.nextFloat();
+                        values[i][j] = random.nextDouble();
                         break;
                 }
             }
